// frontend/src/api/generated/metrics/metrics.ts
export type dashboardMetricsApiV1MetricsDashboardGetResponse200 = {
  data: DashboardMetrics;
  status: 200;
};

export type dashboardMetricsApiV1MetricsDashboardGetResponse422 = {
  data: HTTPValidationError;
  status: 422;
};

export type dashboardMetricsApiV1MetricsDashboardGetResponseSuccess =
  dashboardMetricsApiV1MetricsDashboardGetResponse200 & {
    headers: Headers;
  };
export type dashboardMetricsApiV1MetricsDashboardGetResponseError =
  dashboardMetricsApiV1MetricsDashboardGetResponse422 & {
    headers: Headers;
  };

export type dashboardMetricsApiV1MetricsDashboardGetResponse =
  | dashboardMetricsApiV1MetricsDashboardGetResponseSuccess
  | dashboardMetricsApiV1MetricsDashboardGetResponseError;